Address review: profile registration is the only embedded detection
Removes the handshake env, token env, and port probe fallback
mechanisms along with embedded.go and its tests; the store lookup with
the live embedded metadata gate and the stale-registration note fold
into resolve.go as plain case handling. engine_status drops the source
labeling since only one mechanism exists, and the detection TTL cache
goes with the probes it amortized. Grants, reserved-name filtering,
and denial wording are unchanged.
